LayoutTests/fast/canvas/canvas-strokePath-cap-join-expected.txt

Ensure correct behavior of canvas with path stroke with cap and join

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS successfullyParsed is true

TEST COMPLETE
